requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets==12.0

# Google APIs
google-api-python-client==2.130.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.0

# YouTube
yt-dlp>=2024.12.6

# Web search
requests==2.32.0

# Video processing
ffmpeg-python==0.2.0

# Environment
python-dotenv==1.0.1

# ComfyUI client
aiohttp==3.10.0

# File handling
python-docx==1.1.0
